feat: refactor draggableCourse
  * extracted onClicks to outer componenet
  * now tutorial can be tested withour side effects
